+++ R/LonLat2XY.R
@@ -10,7 +10,7 @@
 #' @param ypix length of tile in pixels
 #' @return a data frame with columns X, Y, x, y
 #' @author David Kahle \email{david@@kahle.io}, based on
-#'   [RgoogleMaps::LatLon2XY()] by Markus Loecher of Sense Networks
+#'   `RgoogleMaps::LatLon2XY()` by Markus Loecher of Sense Networks
 #'   \email{markus@@sensenetworks.com}
 #' @seealso \url{https://wiki.openstreetmap.org/wiki/Slippy_map_tilenames}
 #' @export